Add RoutesParser

+++ feature/sydney-trains/database/api/src/main/kotlin/xyz/ksharma/krail/database/sydney/trains/database/api/RoutesStore.kt
@@ -29,7 +29,7 @@ interface RoutesStore {
         routeShortName: String,
         routeLongName: String,
         routeDesc: String?,
-        routeType: Int,
+        routeType: Long,
         routeUrl: String?,
         routeColor: String?,
         routeTextColor: String?
